services/core-api/requirements.txt
asgiref==3.10.0
django==4.2.25
sqlparse==0.5.3
typing-extensions==4.15.0
tzdata==2025.2
djangorestframework==3.16.1
python-dotenv==1.1.1
